app/src/main/java/com/movies/appmoviescit/utils/MoviesImageBuilder.kt
package com.movies.appmoviescit.utils

import com.movies.appmoviescit.services.MoviesConstants


private val POSTER_URL = "https://image.tmdb.org/t/p/w154"
private val BACKDROP_URL = "https://image.tmdb.org/t/p/w780"

class MoviesImageBuilder {

    fun buildPosterUrl(posterPath: String): String {
        return POSTER_URL + posterPath + "?api_key=" + MoviesConstants.API_KEY
    }

    fun buildBackdropUrl(backdropPath: String): String {
        return BACKDROP_URL + backdropPath + "?api_key=" + MoviesConstants.API_KEY
    }
}
